Extracted RolePlaying to societies subpackage
