+++ camel/societies/role_playing.py
