+++ examples/ai_society/role_playing.py
@@ -13,7 +13,7 @@
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 from colorama import Fore
 
-from camel.agents import RolePlaying
+from camel.societies import RolePlaying
 from camel.utils import print_text_animated
