+++ examples/misalignment/role_playing_multiprocess.py
@@ -17,8 +17,8 @@ import os
 
 from colorama import Fore
 
-from camel.agents import RolePlaying
 from camel.configs import ChatGPTConfig
+from camel.societies import RolePlaying
 from camel.typing import TaskType
